Rename minValue to better reflect fueling goal
`minValue` was introduced as the value in ether the DAO needs to accumulate
in order to be considered fueled. With the extra balance also taking in
ether and not counting towards the `totalSupply` this is inaccurate.

For this reason we rewrite the `minValue` as `minTokensToCreate` and
interpret it as the minimum amount of wei-equivalent tokens that the DAO
needs to create in order to be considered fueled.

There is also a fix for the extra balance fueling test which was failing
1 in 3 times exactly because the `minValue` goal was denominated in wei
and had not realized the `extraBalance` balance does not count towards
that goal