Add refund after extraBalance fueling test
Adding a Test which fails to fuel the DAO while the extraBalance account
is in use. Afterwards we check that refund works and the full amount is
given back to the user including their portion of what should be in the
extraBalance.

The test works since it fails before commit
9ebf64b44bf5948a2df48a7072168106fcb1f69b but after that commit it is
always succesful.