Stop recursive executeProposal() call
Such a scenario would be very easily spotted and stopped
during the security vetting of the recipient contract and such a
contract would never make it in the whitelist.

Nevertheless, in the case of an obviously malicious contract that calls
executeProposal() recursively making into the whitelist, we are moving
the `proposalPassed` guard variable before the call() to the recipient
contract.

This will protect against the recipient contract calling
`executeProposal()` again recursively and effectively emptying the DAO
from funds.

Thanks for spotting this @chriseth!