Add an external voteStatusCheck to PFOffer
This is a port of @avsa's contribution from his proposed offer contract
that can be seen here:
https://gist.github.com/alexvandesande/122c05a0f8a7108523aed29ae7ec92b1

At any point `voteStatusDeadline` seconds before the end of the voting
period anyone can call `checkVoteStatus()` to attempt and set the
`wasApprovedBeforeDeadline` flag.

This allows for a way to assure that the last minute yes vote does not
occur by requiring the flag to be set to true at an earlier point of the
voting deadline.